moved scrape.ts to pages/api directory
